Refine the GDN kernels: relocate files, relocate utility to Utils.h, removed unused utilities
