+++ src/GdnAttnXe20.cmake
@@ -4,5 +4,5 @@
 # live in a subdir and are added to the Xe20 source list explicitly so they
 # compile with -device bmg.
 list(APPEND device_cpp_xe20
-    "${CMAKE_CURRENT_SOURCE_DIR}/sycl/kernels/gdn_attn/xe_2/chunk_gated_delta_rule_xe2.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/sycl/kernels/gdn_attn/xe_2/l2norm.cpp")
+    "${CMAKE_CURRENT_SOURCE_DIR}/sycl/gdn_attn/chunk_gated_delta_rule.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/sycl/gdn_attn/l2norm.cpp")
