+++ src/sycl/Utils.h
@@ -58,6 +58,19 @@ static inline void barrier() {
   asm volatile("barrier\n" ::: "memory");
 }
 
+namespace syclex = sycl::ext::oneapi::experimental;
+
+static inline syclex::architecture get_device_architecture(at::DeviceIndex device_index = -1) {
+  auto device_id = (device_index == -1) ? c10::xpu::current_device() : device_index;
+  auto raw_device = c10::xpu::get_raw_device(device_id);
+  return raw_device.get_info<syclex::info::device::architecture>();
+}
+
+static inline bool is_bmg(at::DeviceIndex device_index = -1) {
+  return get_device_architecture(device_index) == syclex::architecture::intel_gpu_bmg_g21 ||
+         get_device_architecture(device_index) == syclex::architecture::intel_gpu_bmg_g31;
+}
+
 using DeviceId = at::DeviceIndex;
 
 static inline DeviceId dpcppGetDeviceIdOfCurrentQueue() {
@@ -359,4 +372,17 @@ template <>
 struct DtypeInfo<c10::Float8_e4m3fn> {
   static constexpr float MIN = -448;
   static constexpr float MAX = 448;
+};
+
+template <typename scalar_t, int vec_size>
+struct alignas(sizeof(scalar_t) * vec_size) aligned_vec {
+  scalar_t val[vec_size];
+
+  scalar_t& operator[](int index) {
+    return val[index];
+  }
+
+  scalar_t const& operator[](int index) const {
+    return val[index];
+  }
 };
