--- src/sycl/kernels/gdn_attn/xe_2/chunk_gated_delta_rule_xe2.cpp
@@ -1,44 +1,0 @@
-#define SYCL_INTEL_TARGET 20
-#include "chunk_gated_delta_rule_xe2.h"
-
-#include <torch/all.h>
-
-#include <sycl/sycl.hpp>
-
-#include "chunk_gated_delta_rule_kernels_xe2.hpp"
-
-void chunk_gated_delta_rule_xe2(
-    sycl::queue& queue,
-    torch::Tensor& core_attn_out,                           // [total_seqlen, num_v_heads, head_v_dim]
-    const torch::Tensor& q,                                 // [total_virtual_seqlen, num_k_heads, head_k_dim]
-    const torch::Tensor& k,                                 // [total_virtual_seqlen, num_k_heads, head_k_dim]
-    const torch::Tensor& v,                                 // [total_virtual_seqlen, num_v_heads, head_v_dim]
-    const torch::Tensor& b,                                 // [num_v_heads, total_virtual_seqlen]
-    const torch::Tensor& a,                                 // [num_v_heads, total_virtual_seqlen]
-    const torch::Tensor& A_log,                             // [num_v_heads]
-    const torch::Tensor& dt_bias,                           // [num_v_heads]
-    torch::Tensor& ssm_state,                               // [cache_batch_size, num_v_heads, head_v_dim, head_k_dim]
-    const torch::Tensor& query_start_loc,                   // [batch_size + 1]
-    const torch::Tensor& cache_indices,                     // [batch_size]
-    const std::optional<torch::Tensor>& has_initial_state,  // [batch_size] or None
-    const int num_prefills,
-    const int num_decodes,
-    const int* token_indx) {
-  gdn::chunk_gated_delta_rule_impl_xe2(
-      queue,
-      core_attn_out,
-      q,
-      k,
-      v,
-      b,
-      a,
-      A_log,
-      dt_bias,
-      ssm_state,
-      query_start_loc,
-      cache_indices,
-      has_initial_state,
-      num_prefills,
-      num_decodes,
-      token_indx);
-}
