--- src/sycl/kernels/gdn_attn/xe_2/l2norm.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <torch/all.h>
-
-#include <sycl/sycl.hpp>
-
-void l2norm(sycl::queue& queue, const torch::Tensor& q, const torch::Tensor& k);
